--- src/components/Main/Search/styles.js
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-	
-`
-
-export const SearchInput = styled.input`
-	
-`
