--- src/components/Main/styles.js
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-	height: 100%;
-	width: 100%;
-`
-
-export const Title = styled.p`
-	font-size: 24px;
-	padding-bottom: 32px;
- 	text-align: center;
-`
