+++ src/components/index.js
@@ -1,7 +1,19 @@
 import React from "react";
+import styled from "styled-components";
+
 import Search from "./Search";
 import Card from "./Card";
-import { Container, Title } from "./styles";
+
+export const Container = styled.div`
+	height: 100%;
+	width: 100%;
+`
+
+export const Title = styled.p`
+	font-size: 24px;
+	padding-bottom: 32px;
+ 	text-align: center;
+`
 
 const Main = () => (
   <Container>
