--- src/styles/index.js
@@ -1,4 +1,0 @@
-/****
-*	If you are familiar with styled-components, feel free
-*	to add any global styling you'd like here & add it into the project!
-****/
